Add model to app
Added a create_model script, which creates a new model, and updated classifier script to use this model. Still having issues with the input layers
