--- constants.py
@@ -1,2 +1,0 @@
-IMAGE_HEIGHT = 500
-IMAGE_WIDTH = 500
