+++ scripts/constants.py
@@ -1,0 +1,2 @@
+IMAGE_HEIGHT = 500
+IMAGE_WIDTH = 500
